feat(corehandler): Refactor
