+++ cg/store/crud/delete.py
@@ -3,7 +3,7 @@
 
 from sqlalchemy.orm import Session
 
-from cg.store.api.base import BaseHandler
+from cg.store.base import BaseHandler
 from cg.store.filters.status_flow_cell_filters import (
     FlowCellFilter,
     apply_flow_cell_filter,
